Remove type conversion for boolean data

+++ Boolean(Basic)/Visualize_Boolean_Tensor_GL_IMGUI.py
@@ -283,7 +283,7 @@ def show_2d_tensor(tensor):
             array,
             0,
             0,
-            tensor.to(torch.uint8).data_ptr(),
+            tensor.data_ptr(),
             tensorWidth,
             tensorWidth,
             tensorHeight,
